requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
azure-ai-agents==1.2.0b3
azure-ai-inference==1.0.0b9
azure-ai-projects==1.1.0b3
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
descope==1.7.9
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
liccheck==0.9.2
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
narwhals==2.3.0
numpy==2.3.2
openai==1.106.0
opentelemetry-api==1.36.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
semantic-version==2.10.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.49.1
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1